Add 2 live example and update source

+++ my_orb_slam2/include/Map.h
@@ -70,7 +70,7 @@ namespace ORB_SLAM2
 		uint64_t MapPointsInMap();
 		uint64_t KeyFramesInMap();
 
-		uint32_t GetMaxKFid();
+		uint64_t GetMaxKFid();
 
 		void Clear();
